Fix webhook crash on payment failed on unknown sub

+++ src/subscription/subscription.controller.ts
@@ -76,6 +76,9 @@ export class SubscriptionController {
       );
       const subscriptionId = event.data.object['id'];
       const customer = await this.customerService.getBySubId(subscriptionId);
+      if (!customer) {
+        return;
+      }
       this.userService.userUnsubscribed(customer.userId).subscribe();
       await this.subscriptionService.deleteSubscription(subscriptionId);
     } catch (err) {
